ragas_eval/requirements.txt
ragas>=0.1.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0
python-dotenv>=1.0.0
